bhagvadgpt-backend/requirements.txt
fastapi
uvicorn
pydantic
chromadb
sentence-transformers
langchain-groq
langchain-corepip install -r requirements.txt.\venv\Scripts\activate